+++ docs/source/change_log.rst
@@ -22,6 +22,8 @@ Version 1.12.0
    * Plugins can ship with dedicated documentation in markdown files that will be displayed
    * The GUI no longer locks up while tasks like downloading plugins are taking place
 
+* Added a ``campaign-expired`` server signal
+
 Version 1.11.0
 ^^^^^^^^^^^^^^